requirements.txt
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.108.0
# google-generativeai==0.3.2
# python-dotenv==1.0.0
# beautifulsoup4==4.12.2
# lxml==4.9.3
# python-dateutil==2.8.2
# sqlalchemy==2.0.23
# pydantic==2.5.0
# apscheduler==3.10.4
# email-validator==2.1.0

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
python-dotenv
beautifulsoup4
lxml
python-dateutil
sqlalchemy
pydantic
apscheduler
email-validator
flask
schedule



